requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
protobuf==3.19.4
pybind11==2.9.0
pycodestyle==2.8.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
